# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/ingest/vm/findings": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ingest vulnerability findings from VM scanners (Tenable, Qualys, etc.)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingestion"
                ],
                "summary": "Ingest VM findings",
                "parameters": [
                    {
                        "description": "VM findings payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ingest.VMFindingsPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ingest summary",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ingest.InstalledSoftware": {
            "type": "object",
            "properties": {
                "cpe": {
                    "description": "CPE string (if provided by scanner)",
                    "type": "string"
                },
                "edition": {
                    "type": "string"
                },
                "install_path": {
                    "description": "Install location (optional)",
                    "type": "string"
                },
                "product": {
                    "type": "string"
                },
                "vendor": {
                    "description": "Software identification",
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "ingest.VMAsset": {
            "type": "object",
            "properties": {
                "external_ids": {
                    "description": "External IDs (e.g., AWS instance ID, Azure resource ID)",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "hostname": {
                    "description": "Hostname (primary locator)",
                    "type": "string"
                },
                "installed_software": {
                    "description": "Installed software (optional, for software inventory)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ingest.InstalledSoftware"
                    }
                },
                "ip_addresses": {
                    "description": "IP addresses (optional, conditional matching)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "shortname": {
                    "description": "Shortname (optional, for DHCP-heavy environments)",
                    "type": "string"
                },
                "static_ip": {
                    "description": "Whether this asset has a static IP (for conditional IP matching)",
                    "type": "boolean"
                }
            }
        },
        "ingest.VMFinding": {
            "type": "object",
            "required": [
                "asset",
                "finding",
                "last_found",
                "status"
            ],
            "properties": {
                "asset": {
                    "description": "Asset identification",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ingest.VMAsset"
                        }
                    ]
                },
                "evidence": {
                    "description": "Additional evidence",
                    "type": "object",
                    "additionalProperties": true
                },
                "finding": {
                    "description": "Finding details",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ingest.VMFindingDetails"
                        }
                    ]
                },
                "first_found": {
                    "type": "string"
                },
                "last_found": {
                    "type": "string"
                },
                "status": {
                    "description": "Status and timestamps",
                    "type": "string",
                    "enum": [
                        "open",
                        "fixed",
                        "fixed_by_verification"
                    ]
                }
            }
        },
        "ingest.VMFindingDetails": {
            "type": "object",
            "required": [
                "definition_id",
                "severity",
                "title"
            ],
            "properties": {
                "cves": {
                    "description": "CVE IDs associated with this finding",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "definition_id": {
                    "description": "Scanner-specific definition ID",
                    "type": "string"
                },
                "description": {
                    "description": "Description (optional)",
                    "type": "string"
                },
                "references": {
                    "description": "References (URLs, advisories, etc.)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "severity": {
                    "description": "Severity level",
                    "type": "string",
                    "enum": [
                        "Critical",
                        "High",
                        "Medium",
                        "Low",
                        "Info"
                    ]
                },
                "solution": {
                    "description": "Solution text (optional)",
                    "type": "string"
                },
                "title": {
                    "description": "Title/name of the vulnerability",
                    "type": "string"
                }
            }
        },
        "ingest.VMFindingsPayload": {
            "type": "object",
            "required": [
                "findings",
                "scanned_at",
                "source"
            ],
            "properties": {
                "findings": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/ingest.VMFinding"
                    }
                },
                "scanned_at": {
                    "type": "string"
                },
                "scanner": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
...
    }
}